requirements.txt
discord-py-slash-command==1.1.2
discord.py==1.6.0